refactor : 클래스를 세부적으로 분류
